fix: float into double and add static cast for vector2d

+++ include/ECS/Components/BackgroundComponent.hpp
@@ -16,15 +16,13 @@
 namespace ECS::Components {
     /**
      * @brief Component that contains the data needed for a parallax background
-     * @param spriteSheetIndex The index of the sprite sheet to use
-     * Component that contains the data needed for a parallax background
      * @param texturePath Path to the background texture
      * @param frameScale The scale of the background
      * @see SpriteSheetDrawer
      */
     struct BackgroundComponent {
         Texture2D texture;
-        float frameScale = 1.0;
-        float paralaxSpeed = 1.0;
+        double frameScale = 1.0;
+        double paralaxSpeed = 1.0;
     };
 } // namespace ECS::Components
